+++ react/Shelf.js
@@ -1,52 +1,76 @@
 import PropTypes from 'prop-types'
-import { path } from 'ramda'
-import React, { Component } from 'react'
+import { path, identity } from 'ramda'
+import React, { Component, useMemo, useEffect } from 'react'
 import { graphql } from 'react-apollo'
-import { compose, branch, renderComponent } from 'recompose'
-import { withRuntimeContext, Loading } from 'vtex.render-runtime'
+import { withRuntimeContext, Loading, useRuntime } from 'vtex.render-runtime'
+import { usePixel } from 'vtex.pixel-manager/PixelContext'
 
 import OrdenationTypes, {
   getOrdenationNames,
   getOrdenationValues,
-} from './OrdenationTypes'
-import ProductList from './ProductList'
-import { productListSchemaPropTypes } from './propTypes'
+} from './utils/OrdenationTypes'
+import ProductList from './components/ProductList'
+import { productListSchemaPropTypes } from './utils/propTypes'
 import productsQuery from './queries/productsQuery.gql'
-import ShelfContent from './ShelfContent'
+import ShelfContent from './components/ShelfContent'
 
-import shelf from './shelf.css'
+import shelf from './components/shelf.css'
+import { normalizeProduct, normalizeBuyable } from './utils/normalize'
+
+const useProductImpression = (products) => {
+  const { push } = usePixel()
+
+  useEffect(() => {
+    if (!products) {
+      return
+    }
+
+    products.forEach((product, index) => {
+      const normalizedProduct = normalizeProduct(product)
+
+      push({
+        event: 'productImpression',
+        list: 'Shelf',
+        position: index + 1,
+        product: normalizedProduct,
+      })
+    });
+  }, [push, products])
+}
 
 /**
  * Shelf Component. Queries a list of products and shows them.
  */
-class Shelf extends Component {
-  shouldComponentUpdate(nextProps) {
-    return !Boolean(this.props.data.error && nextProps.data.error)
+const Shelf = ({ data, productList = ProductList.defaultProps }) => {
+  const { hints: { mobile }} = useRuntime()
+  const { loading, error, products } = data || {}
+
+
+  const filteredProducts =
+    products && products.map(normalizeBuyable).filter(identity)
+
+  const productListProps = useMemo(() => ({
+    products: filteredProducts,
+    loading: loading,
+    isMobile: mobile,
+    ...productList,
+  }), [filteredProducts, loading, mobile, productList])
+
+  useProductImpression(filteredProducts)
+
+  if (loading) {
+    return <Loading />
   }
 
-  render() {
-    const { data, productList, runtime } = this.props
-    const products = path(['products'], data)
-    const productListProps = {
-      products,
-      loading: data.loading,
-      isMobile: runtime.hints.mobile,
-      ...productList,
-    }
-
-    if (data.error || data.loading) {
-      return null
-    }
-    return (
-      <div className={`${shelf.container} pv4 pb9`}>
-        <ProductList {...productListProps} />
-      </div>
-    )
+  if (error) {
+    return null
   }
-}
 
-Shelf.defaultProps = {
-  productList: ProductList.defaultProps,
+  return (
+    <div className={`${shelf.container} pv4 pb9`}>
+      <ProductList {...productListProps} />
+    </div>
+  )
 }
 
 Shelf.propTypes = {
@@ -83,13 +107,7 @@ const options = {
   }),
 }
 
-const EnhancedShelf = compose(
-  graphql(productsQuery, options),
-  branch(
-    props => props.data && props.data.loading,
-    renderComponent(Loading)
-  )
-)(withRuntimeContext(Shelf))
+const EnhancedShelf = graphql(productsQuery, options)(Shelf)
 
 EnhancedShelf.getSchema = props => {
   return {